More work done with styling
Some more work has been done with styling. Getting closer to the final
product. Added a footer that I forgot to create (HTML and CSS) beforehand.
The quote is going to need some more work before I'm able to get it looking
correct and I still have not fixed the issue with the page not auto growing;